+++ docs/griptape-framework/misc/src/events_chunk_stream.py
@@ -1,5 +1,5 @@
 from griptape.drivers import OpenAiChatPromptDriver
-from griptape.events import BaseChunkEvent, EventBus, EventListener
+from griptape.events import ActionChunkEvent, EventBus, EventListener, TextChunkEvent
 from griptape.structures import Pipeline
 from griptape.tasks import ToolkitTask
 from griptape.tools import PromptSummaryTool, WebScraperTool
@@ -8,7 +8,11 @@ EventBus.add_event_listeners(
     [
         EventListener(
             lambda e: print(str(e), end="", flush=True),
-            event_types=[BaseChunkEvent],
+            event_types=[TextChunkEvent],
+        ),
+        EventListener(
+            lambda e: print(str(e), end="", flush=True),
+            event_types=[ActionChunkEvent],
         ),
     ]
 )