+++ griptape/drivers/__init__.py
@@ -112,6 +112,7 @@ from .event_listener.pusher_event_listener_driver import PusherEventListenerDriv
 from .file_manager.base_file_manager_driver import BaseFileManagerDriver
 from .file_manager.local_file_manager_driver import LocalFileManagerDriver
 from .file_manager.amazon_s3_file_manager_driver import AmazonS3FileManagerDriver
+from .file_manager.griptape_cloud_file_manager_driver import GriptapeCloudFileManagerDriver
 
 from .rerank.base_rerank_driver import BaseRerankDriver
 from .rerank.cohere_rerank_driver import CohereRerankDriver
@@ -230,6 +231,7 @@ __all__ = [
     "BaseFileManagerDriver",
     "LocalFileManagerDriver",
     "AmazonS3FileManagerDriver",
+    "GriptapeCloudFileManagerDriver",
     "BaseRerankDriver",
     "CohereRerankDriver",
     "BaseRulesetDriver",